# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Simple microservice for borrowing books with jwt auth",
        "title": "Book microservice",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Dias Galikhanov",
            "email": "diasgalikhanov@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8586",
    "basePath": "/api/v1",
    "paths": {
...
        "/sign-in": {
            "post": {
                "description": "login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "SignIn",
                "operationId": "login",
                "parameters": [
                    {
                        "description": "credentials",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AuthUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "token",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "default": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/sign-up": {
            "post": {
                "description": "create account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "SignUp",
                "operationId": "create-account",
                "parameters": [
                    {
                        "description": "account info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "id",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "default": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/update-password": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "update password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "UpdatePassword",
                "operationId": "update-password",
                "parameters": [
                    {
                        "description": "current and new password",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdatePasswordReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "default": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/users-with-books": {
            "get": {
                "description": "get users with books they borrowed",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "GetUsersWithBooks",
                "operationId": "get-users-with-books",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.UserBook"
                            }
                        }
                    },
                    "400": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "default": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/users-with-books-for-month": {
            "get": {
                "description": "get users with books they borrowed for last month",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "GetUsersWithBooksForMonth",
                "operationId": "get-users-with-books-for-month",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.UserBook"
                            }
                        }
                    },
                    "400": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "default": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.AuthUser": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "models.Book": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "models.UpdatePasswordReq": {
            "type": "object",
            "properties": {
                "current_password": {
                    "type": "string"
                },
                "new_password": {
                    "type": "string"
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "user_role": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "models.UserBook": {
            "type": "object",
            "properties": {
                "books": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Book"
                    }
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}